--- frontend/techpick-extension/src/utils/index.ts
@@ -1,9 +1,0 @@
-export { handleHTTPError } from './handleHTTPError';
-export { returnErrorFromHTTPError } from './returnErrorFromHTTPError';
-export { hasIndex } from './hasIndex';
-export { numberToRandomColor } from './numberToRandomColor';
-export { isShallowEqualValue } from './isShallowEqualValue';
-export { isEmptyString } from './isEmptyString';
-export { getElementById } from './getElementById';
-export { extractOpenGraphMetadata } from './extractOpenGraphMetadata';
-export { filterSelectableFolder } from './filterSelectableFolderList';
